+++ tests/ResponseTest.php
@@ -193,6 +193,8 @@ class ResponseTest extends TestCase
 
     /**
      * Test setExpiry method with null expiry time value.
+     *
+     * @noinspection PhpDeprecationInspection
      */
     public function testSetExpiryWithNullTime()
     {
@@ -205,6 +207,8 @@ class ResponseTest extends TestCase
 
     /**
      * Test setExpiry method with past expiry time value.
+     *
+     * @noinspection PhpDeprecationInspection
      */
     public function testSetExpiryWithPastTime()
     {
@@ -219,6 +223,8 @@ class ResponseTest extends TestCase
 
     /**
      * Test setExpiry method with future expiry time value.
+     *
+     * @noinspection PhpDeprecationInspection
      */
     public function testSetExpiryWithFutureTime()
     {
@@ -228,6 +234,46 @@ class ResponseTest extends TestCase
         $response->setExpiry($expiry);
 
         self::assertSame($expiry->setTimezone(new DateTimeZone('UTC'))->format('D, d M Y H:i:s \G\M\T'), $response->getHeader('Expires'));
+        self::assertSame('public, max-age=86400', $response->getHeader('Cache-Control'));
+    }
+
+    /**
+     * Test setExpiryDateTime method with null expiry time value.
+     */
+    public function testSetExpiryDateTimeWithNullTime()
+    {
+        $response = new Response();
+        $response->setExpiryDateTime(null);
+
+        self::assertSame($response->getHeader('Date'), $response->getHeader('Expires'));
+        self::assertSame('no-cache, no-store, must-revalidate, max-age=0', $response->getHeader('Cache-Control'));
+    }
+
+    /**
+     * Test setExpiryDateTime method with past expiry time value.
+     */
+    public function testSetExpiryDateTimeWithPastTime()
+    {
+        $response = new Response();
+
+        $expiryDateTime = (new DateTimeImmutable())->sub(new DateInterval('PT24H'));
+        $response->setExpiryDateTime($expiryDateTime);
+
+        self::assertSame($expiryDateTime->setTimezone(new DateTimeZone('UTC'))->format('D, d M Y H:i:s \G\M\T'), $response->getHeader('Expires'));
+        self::assertSame('no-cache, no-store, must-revalidate, max-age=0', $response->getHeader('Cache-Control'));
+    }
+
+    /**
+     * Test setExpiryDateTime method with future expiry time value.
+     */
+    public function testSetExpiryDateTimeWithFutureTime()
+    {
+        $response = new Response();
+
+        $expiryDateTime = (new DateTimeImmutable())->add(new DateInterval('PT24H'));
+        $response->setExpiryDateTime($expiryDateTime);
+
+        self::assertSame($expiryDateTime->setTimezone(new DateTimeZone('UTC'))->format('D, d M Y H:i:s \G\M\T'), $response->getHeader('Expires'));
         self::assertSame('public, max-age=86400', $response->getHeader('Cache-Control'));
     }